requirements.txt
Django==2.0
gunicorn==19.6.0
whitenoise==3.2.1
dj-database-url==0.4.1
djangorestframework
django-rest-framework-social-oauth2==1.0.4
django-bootstrap3==7.0.1
celery
django-google-analytics-app
reportlab==3.2.0

click==6.7
colorama==0.3.9
jinja2==2.10
tornado==4.5.2
tornado-jinja2==0.2.4
pyhocon==0.3.38
structlog[dev]==17.2.0
ujson==1.35
uvloop==0.13.0
watchdog==0.8.3
async_lru==1.0.1